fix: pr feedback and failing tests
