+++ package.json
@@ -49,7 +49,7 @@
     "jest-styled-components": "^7.0.2",
     "koa": "^2.8.0",
     "nodemon": "^2.0.15",
-    "strapi-typed": "1.0.10",
+    "strapi-typed": "1.0.13",
     "ts-jest": "^27.1.3",
     "ts-node": "^10.7.0",
     "typescript": "^4.5.5"
